requirements.txt
blinker==1.9.0
cachetools==7.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
et-xmlfile==2.0.0
flask==3.1.3
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
openpyxl==3.1.5
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.6.3
werkzeug==3.1.5
